style(sources): rename defs.h to wx_common.h
