+++ src/attachmentdialog.h
@@ -19,7 +19,7 @@
 #ifndef MM_EX_ATTACHMENTDIALOG_H_
 #define MM_EX_ATTACHMENTDIALOG_H_
 
-#include "defs.h"
+#include "wx_common.h"
 #include <wx/dataview.h>
 #include <map>
