+++ src/import_export/qif_import.h
@@ -19,7 +19,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #ifndef QIF_IMPORT_H
 #define QIF_IMPORT_H
 
-#include "defs.h"
+#include "wx_common.h"
 #include <vector>
 
 // http://en.wikipedia.org/wiki/QIF
