+++ src/mmcustomdata.h
@@ -18,7 +18,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 ********************************************************/
 
 #pragma once
-#include "defs.h"
+#include "wx_common.h"
 #include "Model_CustomField.h"
 #include "Model_Checking.h"
