+++ src/optiondialog.h
@@ -21,7 +21,7 @@
 #ifndef MM_EX_OPTIONSDIALOG_H_
 #define MM_EX_OPTIONSDIALOG_H_
 
-#include "defs.h"
+#include "wx_common.h"
 #include <vector>
 
 class wxListbook;
