+++ src/splitdetailsdialog.h
@@ -20,7 +20,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #ifndef MM_EX_SPLITDETAILSDIALOG_H_
 #define MM_EX_SPLITDETAILSDIALOG_H_
 
-#include "defs.h"
+#include "wx_common.h"
 #include "mmTextCtrl.h"
 #include "Model_Splittransaction.h"
