+++ src/webappdialog.h
@@ -19,7 +19,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #ifndef MM_EX_WEBAPPDIALOG_H_
 #define MM_EX_WEBAPPDIALOG_H_
 
-#include "defs.h"
+#include "wx_common.h"
 #include "webapp.h"
 #include <wx/dataview.h>
 #include <wx/srchctrl.h>
