+++ src/wx_common.h
